Move SyncState to sync.ts
